% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils_io.R
\name{ReadGraphFile}
\alias{ReadGraphFile}
\title{Read graph file}
\usage{
ReadGraphFile(fileName, fileType, org, idType)
}
\arguments{
\item{fileName}{file name of the graph file}

\item{fileType}{file type of the graph file (graphml, sif, txt, json)}

\item{org}{(optional, can be NA) organism}

\item{idType}{(optional, can be NA) identifier type}
}
\description{
read graph file and store into graph object
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
